+++ samp/src/lib.rs
@@ -62,6 +62,7 @@ pub mod interlayer;
 pub mod logger;
 #[cfg(not(feature = "samp-only"))]
 pub(crate) mod macros;
+pub mod mainthread;
 pub mod plugin;
 pub(crate) mod runtime;